Always make space for nullsBuffer and hashTablesBuffer
Block.getSizeInBytes() and Block.getLogicalSizeInBytes() always adds up
the sizes of nulls buffer even if the block cannot contain nulls. When
estimating the max buffer capacity for BlockEncodingBuffers, we can also
leave the space for the nullsBuffer and hashTablesBuffer. This will not
waste memory because the buffers are not actually allocated until blocks
with nulls or hashtables come in. It will make the buffers sizes
proportional to the blocks' logical sizes, and make the code cleaner.