Invoke CBO at SqlQueryScheduler for Join Swapping
- Pass in CBO and make it invokable at [Legacy/]SqlQueryScheduler
during runtime.
- Rename get() method of PlanOptimizers to
getPlanningTimeOptimizers() to distinguish getRuntimeOptimizers()
- Adjust IterativeOptimizer optimize() function to return the original
plan when the plan is not changed, instead of always calling
memo.extract() to rebuild a new one.
- Create a join swapping rule (RuntimeReorderJoinSides) based on
the probe and build side statistics, and adjust the local exchange
when necessary (add at build side, remove at probe side).
- Rebuild the section, re-generate stageExecutionAndSchedulers of the
section, and adjust the overall subplan when the join is swapped to
reflect the correct statistics to web UI and QueryCompletionEvent.