Use the same RowExpressionCompiler in CursorProcessorCompiler
In CursorProcessorCompiler, we created new RowExpressionCompiler when compiling
every expression. Now that RowExpressionCompiler is no longer stateless
(compiledLambdaMap needs to be updated), this becomes problematic. Conceptually,
RowExpressionCompiler is adding methods to a generated class, and all codegen
related to the same generated class should use the same RowExpressionCompiler.
Modify the code accordingly.