Fix For Issue #14269
Linked Issue:
https://github.com/prestodb/presto/issues/14269

When instantiating a PrestoPreparedStatement class object with a limit argument containing an '?' parameter, the constructor runs a super PrestoStatement execute function on the full PREPARE sql query you're instantiating with. This means that the PREPARE statement is compiled/parsed through ANTLR, hence the "org.antlr.v4.runtime" classpath in the above stack trace. When ANTLR hits the SELECT limit argument, it checks its lexer rule for limit and it doesn't contain the '?' parameter character as a listed, valid token.

When instantiating a PrestoPreparedStatement, the class should not execute the statement in the constructor. This changes the behavior inherited from the parent PrestoStatement class and breaks prepared statements using sql parameters.