Avoid fetching encryption information when no columns requested
This can happen if the user is only asking for partition columns or something like count(*). As of now,
if the table has table level encryption key, we will try getting the key metadata for that key where
as it is not needed. This change will allow users who don't have access to the table data to still be able
to get the partition columns or count(*) and bring the implementation at par with column level encryption
key specification