Allow EnforceSingleRowNode for predicate extractor
Assuming we have a daily ingested table that is partitioned on ds,
a filter like `ds = (SELECT '2020-07-01')` is converted into an
INNER JOIN, but this value is not passed to the other side of Join,
which leads to full table scan.

This commit will enable this value being treated as predicate, and
thus we only need to read this one partition.