Support evaluating min/max only metadata query
Assuming we have a daily ingested table that is partitioned on ds, one
common use case is to fetch data from latest ds partition. One way to
compose such a query is using a filter like
`ds = (SELECT max(ds) FROM table)`. However, this filter is converted
into an INNER JOIN, and will lead to a full table scan on the other
side of join.

Instead, this commit enables a query like `SELECT max(ds) FROM table`
being evaluated at optimization time when OPTIMIZE_METADATA_QUERIES
is set to true, and convert it into a ValuesNode, which could then
be pushed to the other side of Join to avoid expensive full table
scan.