Honor ignore_stats_calculator_failures in getTableStatistics
ignore_stats_calculator_failures was getting enforced in
cachingStatsProvider and cachingCostProvider.  However, InputExtractor
was calling getTableStatistics directly, so queries could still fail due
to errors computing stats.  This was particularly a problem when filters
were pushed down to the table scan, because we compute filter stats as
part of getTableStatistics().