Fix failure during optimization of LIKE expression
Stats calculation code was calling optimize on the like expression
twice.  This was causing query failures, since the second time it was
called, the second like argument had already been converted to a regex.
Instead, just mark the expression as unchanged if it's no longer a like
pattern function or cast function.