Support FilterNode on top of AggregationNode
"having" condition for sql involves "group by" is common case, and it
will generate a FilterNode on top of AggregationNode. Previously,
The SQL will fail directly, and now it allows such case. However, such
filter is not pushed down to Druid, just filter at presto side because
we could not confirm all filters on top of AggregationNode have
corresponding 'dql' pattern to describe it.