Support distributed and local UNION ALL in Presto-on-Spark
Boadcasted UNION ALL is not distributed but collected on a single node
because PlanFragmenter does not allow multiple table scans. This PR
fixes it by changing query planner to support local UNION ALL in a
single fragment if possible. Fragment with colocated join or aggregation
is scheduled separately as before.