RowGroup pruning using the filter on pushed down subfield
Once the dereference is pushed down to tablescan, filters on the dereference column
are automatically pushed down to tablescan as the pushed down dereference column is
exposed as a column directly coming from the tablescan. In the reader (this case
Parquet), interpret pushed down dereference columns and convert them to column paths
to look up the stats in footer.

Co-authored-by: Zhenxiao Luo <zluo@twitter.com>
Co-authored-by: qqibrow <qqibrow@gmail.com>