Fix pruning unreferenced variables when WindowNode is skipped in PruneUnreferencedOutputs
Part of the a query got into the following plan:

```
  Project(user_uuid_506)
    Filter (rank_554 = 1)
      Window(rank_554=rank(), orderingScheme=expr_538,random_553)
        Project(user_uuid_506, expr_538, random_553=random())
          Values(user_uuid_506, expr_538(cast(name as date)))
```

As the Values() operator doesn't have any data (original table got pruned), the Filter is removed as part of the PredicatePushdown rule

```
  Project(user_uuid_506)
    Window(rank_554=rank(), orderingScheme=expr_538,random_553)
      Project(user_uuid_506, expr_538, random_553=random())
        Values(user_uuid_506, expr_538(cast(name as date)))
```

In the remaining plan when trying to prune the unreferenced outputs (PruneUnreferencedOutputs rule),
we remove the WindowNode as no one is looking at the rank_554, but the unreferenced variables expr_538
and random_553 are not removed as we pass the context that has variables from `Project(user_uuid_506)`
and `Window(rank_554=rank(), orderingScheme=expr_538,random_553)`. This causes the issues later on
in PushdownSubfields rule which expects the plan to have unreferenced output variables pruned.

One another way to resolve this is add one more PruneUnreferencedVariables rule just before the
PushdownSubfields rule, but it is better to fix this rule so that it always removes unreferenced
variables for the first time.