Create OrcPageSource using the OrcDataSource from OrcReader
OrcReader would wrap the OrcDataSource with CachingOrcDataSource if the
file is less than the tiny stripe threashold. However the PageSource in
the scan operators would still use the original OrcDataSource. This
is confusing since the OrcReader, OrcRecordReder, and StreamReaders all
keep the CachingOrcDataSource while the PageSource keeps the underlying
OrcDataSource. This commit creates the PageSource using the wrapped
CachingOrcDataSource if the original OrcDataSource was wrapped.