Fix page splitter creating large dictionary page list
Page splitter does not work well for dictionary blocks. When a split
happens on a dictionary block, the dictionary will be shallow copied
into the two split pages. Both these two pages will still be large
because of the dictionary. Recursion will continue until there is one
row per page and a large dictionary in each split page. The fix is to
stop splitting once we do not see enough improvement after a split.