Support explain verification
Add a new functionality to the verifier to allow explain verification.

For each query pairs to be verified, explain both control and test, and
compare the query plans. Verification is marked as succeeded as long as
the test query can be explained.

Introduced a new field matchType in the output event, which can be used
to indicate whether there is plan difference.

For non-DML queries, we don't care much about plan difference and hence
the control query and plan comparison are skipped.