Fix IllegalArgumentException in RelationPlanner
in RelationPlanner, when we process values node, we would add coercions for
type only coercion expression. This could change the expression, which does
not exist in the analysis. For Enum literals, we want to rewrite the
dereference to enum literal. Previously this is done after the coercion rewrite.
Thus if the original expression has type only coercions, we could get
IllegalArgumentException when trying to get the type of the node. Moving the
dereference to enum rewrite before the coercion should solve this problem.