Consider predicate columns for encryption
In encryption we pass the list of columns that are being read to help with
getting the credentials only for those columns and allow for more granular
access control.

Uptil now, only the columns being requested in the select statement were
the only ones being considered. This commit adds support for columns in
the predicate as well. We now consider the predicate columns as well.

A special case to callout is `ROW`. Presto currently supports encryption
granularity of a subfield, however, it is non-trivial to get the exact
subfields being accessed for a `ROW`. This PR puts the whole `ROW` column
if any subfield is being used in the predicate. This is a limitation for
now.