Fix potential deadlock when PageBufferClient exceeds memory limit
Prior to this change, if the PageBufferClient caused a task to exceed
its memory limit inside of addPages, the exception would bubble to the
root of the exchange client callback executor without ever marking the
task as failed. If no other operator also saw the memory limit
exceeded condition, then the task would become stuck in a deadlocked
state.