Remove BlockEncodingManager's dependency on TypeManager
BlockEncodingManager as a low level class manager block serde should not depend
on type systems. The dependency was introduced to get method handles to build
hash tables for map block. However map block hashtables are lazily built so they
are not really needed to encode / decode the map block. Changing the code to
require functionalities needing the hashtable to pass in method handles instead
so we can remove BlockEncodingManager's dependency on TypeManager.