Remove TypeManager.resolveOperator
This API is a hack. TypeManager.resolveOperator internally will call
FunctionManager.resolveOperator. However, FunctionManager is not initialized
in TypeRegistry constructor, but set in FunctionManager's constructo, due to
cyclic dependency. So to be able to call this API, we see following hacky code
in test everywhere:
    TypeManager typeManager = new TypeRegistr();
    // associate typeManager with a function manager
    new FunctionManager(typeManager, ....);

The only reason that this API exists is because MapParametricType.createType
needs to use it. So we did
1) adding TypeManager to ParamatricType.createType
2) adding TypeManager.resolveOperator

just so MapParametricType could provide MethodHandle that would be used to
build hashtable in Map block, while no other API needs
TypeManager.resolveOperator, and no other ParametricType needs access to
TypeManager. Another problem is that the API directly returns a MethodHandle,
which means this API will not work once we support user defined types with
operator implementation that's not a Java MethodHandle.

Since the solution will not work for future generalized caase, removing it for
now, and directly check for MapParametricType in TypeRegistry to special handle
map types.