Add option to do more precise ACL checks
Before this change, Presto would check for column access permission on
all columns referenced in any part of the query.
This behavior can sometimes be undesirable, for example, in this query:

``SELECT name FROM (SELECT * FROM nation)``

During execution of this query, access checks would be performed on all
columns in the table nation, even though only the column ``name`` would
actually be read during the execution of the query, and the other
columns in the table have no impact on the query results.

This change introduces a new sesion property,
``check_access_control_on_utilized_columns_only``, which, when
enabled, will only perform access control checks on columns that would
actually be required to produce the query output, ignoring columns that
are referenced in the query, but are not required to compute the query
results.