Enable reading of FileStatistics in presto-orc
This change enables reading of FileStatistics from the file footer
of encrypted and unencrypted files.

Decrypted FileStatistics are only available in the RecordReaders because
they receive intermediate keys as a c-tor argument and can decrypt the
stats. OrcReader returns plaintext FileStatistics from the file footer
and doesn't attempt to decrypt them. This is the reason why I had to modify
the validation process in the OrcWriteValidation.