Make encryption key explicit when reading DWRF stripes
DWRF StripeInformation only contains keymetadata if the keymetadata for
this stripe would be different from the previous stripe.  However, when
we read stripes in the record reader, we exclude stripes that are not
part of the split, so we may not have info about what the keys were in
the previous stripe.  Therefore, we explicitly set the implied key
information.