Make query result HTTP compression client configurable
By default, presto will GZIP query result JSON payloads sent to the
client. However, especially when the client is connected to the
coordinator over localhost, the added overhead of compressing the
response and then uncompressing it on the client is a losing
proposition.

For queries that are bound only by result processing throughput (eg:
SELECT * FROM <large table>) execution time can reduced by 20-50%
when submitted over a localhost connection with compression disabled.