Add pre-read hueristic to parquet MetadataReader
Attempts to reduce the number of reads required for parquet footers
by pre-reading up to 16KiB from the end of the file. If the metadata
section ends up being < 16KiB long, then one additional round trip
has been saved. When the metadata section is longer, the additional
read occurs to load the remaining portion of the metadata section.