Do not cache buffers in OrcInputStream
Cached buffers tend to expand over time and retain extra memory.
In the situation when the number of ORC streams is high and the data
is very unevenly distributed the overhead of retaining expanded buffer
can be very high.

Allocating buffers is not ideal, as it introduces additional GC overhead,
however the underlying buffers in the OrcDataSourceInput are also allocated,
so the extra overhead introduced by this allocation shouldn't cause any major
issues.

Long term the OrcDataSourceInput should be refactored, and should return a
Slice directly instead of returning a SliceInput.