OrcWriter close errors are masked
If there are errors in the ColumnWriter close, it is masked by
reset checkState
https://github.com/prestodb/presto/blob/master/presto-orc/src/main/java/com/facebook/presto/orc/writer/SliceDictionaryColumnWriter.java#L559

If the ColumnWriters are not closed, reset should not be called, otherwise
it will mask the real error and it will always fail with the IllegalStateException.

Split the code into close and post-close, so that close can be moved outside of the
try block.

It is impossible to write test for this use case, unless the column writers are
refactored into the ColumnWriterFactory and the factory is pluggable.

This regression was introduced in https://github.com/prestodb/presto/pull/13508