Improve error handling for spill
checkSuccess wrapped exceptions in an IllegalArgumentException, which
meant the actual error was buried in the stacktrace. getFutureValue
wrapped the source exception in a runtime exception, but only included
the error stacktrace and not the stack to the  getFutureValue call.
Without the other stack trace, you don't know where in the operator
execution spill failed, which can make debugging harder.

Test Plan: Ran tests with max spilled bytes set to 1B and looked at the
errors.