Implement storage based broadcast join in Presto on Spark
Spark provided broadcast variables are not scalable enough to reliably distribute large volumes of data (gigabytes).
Also broadcast variables cause additional memory overhead, as serialized blocks have to be stored in memory for the
torrent algorithm to function.

This implementation uses distributed storage as a medium. It stores the broadcast data into a distributed storage
and then broadcasts only the pointers (usually file names) with a Broadcast variable to let the executors know where
to read the broadcasted data.

This PR doesn't provide a StorageService implementation. Testing is done in a local mode with a local file system as a
local storage.