Instrument the Orc ReadType
OrcReader issues Read to the storage for several reasons.
1. Reading the tail (PostScript + FileFooter)
2. Reading the FileFooter (if Tail did not completely fetch the FileFooter)
3. Reading the StripeFooter.
4. Reading the Stripe Streams.
5. Reading the header if the file failed to parse.

We want to collect metrics on how much each IO is issued for each type and
optimize the IO parameters for our use case.

Also OrcDataSource now contains total IO issued (readCount)
in addition to the readBytes and readTimeNanos.

For my use case, I plan to implement the derived class of OrcDataSource and capture
this metric and Ultimately I want to reduce the number of IOs and increase the size
of each IO.

I added tests in the TestCachingOrcDataSource, but other than than this change
is just plumbing the readType from the source of the read to the OrcDataSource.