Allow adjusting memory limits with session properties
Allow adjusting memory limits with session properties for Presto on Spark

In Spark hard memory limits are enforced on the container level and can
adjusted by the end user. Since the end user is free to adjust the container
size it should be also allowed to adjust the soft memory limits enforced by
the Presto eval engine.