Limit the codegen class name length
We use the function signature as class name for codegened classes.
When ROW types are used, sometimes the name gets extremely long. Since
we use an AtomicLong as part of the generated class name we should be
able to guarantee uniqueness without using the full name. So truncate
at 100 characters.