Generate shorter aggregation function names
We use return type an input types when generating aggregation function names.
This could make the name really long (tens of KB) when complex ROW types are
used. This would waste memory and make the generated name unreadable. So only
keep the first 10 characters of each type parameter when generating the name.