Reject @BeforeMethod without singleThreaded=true
@BeforeMethod in a class that is not @Test(singleThreaded=true) is
usually (always?) a bug. We should automatically find and reject them.

This change uses Regex negative LookBehind match to find @BeforeMethod
in a class that is not @Test(singleThreaded=true). Only top 200 lines
of a class is scanned to save time and avoid potential stack overflow
if a file is too big. This works because @BeforeMethod should appear at
top a class definition.

Some violations have been caught and are fixed in this change too.