Allow ExchangeClient to notify blocked callers when a threshold buffer size is reached
Currently, blocked callers are notified when a single page arrives into the ExchangeClient.
For high QPS queries that have small amounts of data being fetched, this unintentionally
introduces latency by adding an additional hop: the first page received is the last page
to be sent to the client, but the coordinator is not aware of this until some small time later.
By making the exchange client block on a target result size, we can instead be notified
when the query is completed, or, worst case, when the target result size threshold is
reached.