Fix memory leak in NodeTaskMap
This fixes a leak introduced recently in 26a373c. `memoryUsageTracker` and `cpuUtilizationPercentageTracker`
were being passed as both the referrant and as a part of the cleanup function. This caused 2 references to
these objects - one from the `FinalizerService#finalizers` and `FinalizerService#FinalizeReference#cleanup`
and prevented a cleanup.

Fixing this to just pass all 3 cleanups in a single function along with the `NodeStatsTracker` object.