Use node location instead of nodeId to validate nodes
Currently, before registering a node with DiscoveryNodeManager,
we check if it is allowed to register by sending its nodeId
to NodeStatusService. But with this commit, we will send
nodeLocation instead of nodeId. So from now on, the NodeStatusService is
expected to validate a node based on its nodeLocation instead of nodeId.