Optimize splits serialization for Presto on Spark
Presto on Spark has to enumerate all splits in advance to bind them
to Spark tasks as Spark requires all the splits to be assigned in advance
for failure recovery reasons.

Enumerating all splits in advance could be quite memory intensive,
especially when scanning large datasets.

This PR tries to reduce the memory footprint generated by the split
enumeration.

There are 3 different issues this PR is addressing:

1. It reorganizes the split serialization loop to avoid retaining both,
   serialized and deserialized splits in memory
2. It replaces a text based Json encoding with a Json compatible binary
   Smile encoding.
3. It uses less memory when compressing serialized splits. Instead of
   serializing into a byte array first before compressing the serialization
   is done directly into a compression stream. It reduces memory utilization
   spikes and helps us avoid hitting the maximum byte array length
   restriction that the previous implementation used to hit for certain
   queries.