Add support for passthrough Elasticsearch queries
Cherry-pick of https://github.com/trinodb/trino/pull/3735
This allows running queries over the results of a raw Elasticsearch query.
It extends the syntax of the enhanced ES table names with the following:

SELECT * FROM es.default."<index>$query:<base32-encoded ES query>"

The query is base32-encoded to avoid having to deal with escaping quotes and case
sensitivity issues in table identifiers.

The result of these query tables is a table with a single row and a single column
named "result" of type JSON.

Co-authored-by: Martin Traverso <mtraverso@gmail.com>
Co-authored-by: Manfred Moser <manfred@simpligility.com>