Register accumulators as unnamed
It will remove them from the UI and reduce the memory footprint

If an accumulator is named (the name is assigned) it will be displayed
on the Spark web interface. Spark implements that by calling toString on
the accumulator after each update and storing it in the history. Calling
toString n^2 number of times on the list accumulator and storing the
result takes quadratic memory to the number of tasks executed. Making
the accumulators unnamed removes them from the UI so Spark doesn't have
to store the string snapshots inmemory.