Include all tables in the column mapping of materialized views
Support the computation of 1-to-N column mappings from the view
to the base table columns for materialized views. Matched base
table columns are derived by parsing equi join and set operation.

The computation logic is put into MaterializedViewPlanValidator
to avoid logic duplication. The assumption is a valid materialized
view plan means we are able to derive the right column mappings
for SELECT and REFRESH.